src/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
aiortc==1.5.0
aiohttp==3.9.1
websockets==12.0
pyyaml==6.0.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiosqlite==0.19.0
httpx==0.25.2
redis==5.0.1
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3
openai==1.3.5
python-logging-loki==0.3.1
prometheus-client==0.19.0 